CParamsSet made clonable. Small corrections in CProcessExecutorComp.
git-svn-id: http://b035a0a.online-server.cloud/svn/Acf@1495 5144ca80-2444-0410-90ef-8152f95a0219

+++ Include/iqt/CProcessExecutorComp.cpp
@@ -69,7 +69,13 @@ int CProcessExecutorComp::Execute(const istd::CString& executablePath, const ist
 
 void CProcessExecutorComp::OnComponentDestroyed()
 {
-	m_applicationProcess.kill();
+	if (m_applicationProcess.state() == QProcess::Running){
+		m_applicationProcess.terminate();
+	}
+
+	if (!m_applicationProcess.waitForFinished()){
+		m_applicationProcess.kill();
+	}
 
 	BaseClass::OnComponentDestroyed();
 }
